refactor(vaultregistry): split registry by responsibility
